+++ spec/vizcore/cli_spec.rb
@@ -186,6 +186,7 @@ RSpec.describe Vizcore::CLI do
           "5",
           "--noise-gate",
           "0.03",
+          "--no-reload",
           "--projector"
         ]
       )
@@ -195,6 +196,7 @@ RSpec.describe Vizcore::CLI do
         expect(config.audio_file.to_s).to end_with("spec/fixtures/audio/pulse16_mono.wav")
         expect(config.audio_device).to eq("5")
         expect(config.noise_gate).to eq(0.03)
+        expect(config.reload?).to eq(false)
         expect(config.projector_mode).to eq(true)
       end
       expect(runner).to have_received(:run)